feat: render request params as tables
Request parameters are rendered as a markdown table.

If a body param is specified and it has a schema that doesn’t just refer to a model, then a _Request Body_ section is added where the body schema is listed in details.

The parameters table can include the following columns: in, name, type, required, description, pattern, range, default, unique, multiple_of.

The type column includes format and enum values. If the type is “array” then it can also include type info about array items, e.g. `array of integer`.

The range column will include min and max. This can refer to a numeric value, string length or number of items in an array.

Closes #54