feat: include type of range in range format
When a parameter is limited to a min/max value, a range column is included. Min and max can refer to a numeric param, the length of a string param or the number of items in an array param.

To make this less confusing the range is now formated like this:

```
{min} <= number <= {max}
{min} <= length <= {max}
{min} <= items <= {max}
```

Only one has to be specified
```
{min} <= number
number <= {max}
```

Numbers can be exclusive
```
{min} < number < {max}
```