chore: $ref printed in full, better numbers
BREAKING CHANGE: `$ref` is printed like `$ref: \`${$ref}\``. Most refs should be resolved but in the case a ref is not resolved, it's better to print the entire URI. Number types are printed a bit less redundantly, instead of saying `integer, int32` it will just say `int32`. When numbers are used in ranges it might say `0 < int32 <= 200`.